RE2 port + staged version, cleanup
