+++ oregano/shared/src/main/scala/oregano/internal/Pattern.scala
@@ -68,11 +68,13 @@ object Pattern {
         val (p, g) = compile(r, nextGroup)
         (Pattern.Rep0(p), g)
 
-      // case Regex.Capture(r) => compile(r, nextGroup)
       case Regex.Capture(r) =>
         val groupId = nextGroup
         val (p, g2) = compile(r, nextGroup + 1)
         (Pattern.Capture(groupId, p), g2)
+
+      case _ =>
+        throw IllegalArgumentException(s"Unsupported regex: $regex")
     }
 
   def compile(regex: String): (Pattern, Int) =
